fix: use usize for timespan

+++ sdk/src/assertions/soft_binding.rs
@@ -82,10 +82,10 @@ pub struct SoftBindingScopeMap {
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 pub struct SoftBindingTimespanMap {
     /// Start of the time range (as milliseconds from media start) over which the soft binding value has been computed.
-    pub start: u32,
+    pub start: usize,
 
     /// End of the time range (as milliseconds from media start) over which the soft binding value has been computed.
-    pub end: u32,
+    pub end: usize,
 }
 
 impl SoftBinding {
